Parse time with en_US locale, set localized format with timeStyle
If running this with another locale than en_US, the parsing can fail. Then
when creating the localized time, you can't set the format to "h:mm a"
because that is not how a short time string is represented in many locales
(e.g it would produce "9:41 fm" in Swedish when it should really be just
"9:41"). Setting the timeStyle instead should take care of that.