Added ability to set a string as clock value. Also better German localization.
In many cases, you want to create screenshots for different languages.

One possible solution is to change the language of the simulator.
Another way is to use a scheme that sets a certain launch argument (as
described in this post:
http://useyourloaf.com/blog/2013/07/22/using-launch-arguments-to-test-lo
calizations.html ).

The second method is way faster and the approach that I prefer,
due to the fact, that it does not require the simulator to re-load
springboard to change the entire system language.

Now, the string "09:41" is not accurate (for instance) in German (we
use "09:41") so I want to use my own time format.

I extended the project in a way you can set the value you like without
having to wait until the simulator has changed the system language (not
to mention the clicks needed to get into the language menu).