Add BITCODE_GENERATION_MODE user defined setting. This ensures that generated framework will be correctly built with bitcode for archiving and submitting to Apple. Source: https://medium.com/@heitorburger/static-libraries-frameworks-and-bitcode-6d8f784478a9 Source: http://stackoverflow.com/a/34965178
This change is necessary when creating a Release build of SimulatorStatusMagiciOS.framework, then adding manually to a project (not via Carthage or CocoaPods).
In order to use the framework in a Debug build (ie for automation of screenshots), the framework must be linked and also embedded within project.
If bitcode is enabled for the project, all embedded frameworks must also contain bitcode, hence the requirement that SimulatorStatusMagiciOS.framework generates bitcode.
Note: SimulatorStatusMagiciOS.framework is only invoked in Debug build, not Release build, but xcode currently has no built-in method for embedding frameworks based on configuration.
Source: https://forums.developer.apple.com/thread/95424